move mscw_ctm to own plugin submodule
