+++ pyaerocom/io/__init__.py
@@ -72,7 +72,6 @@ from .read_eea_aqerep_v2 import ReadEEAAQEREP_V2
 from .read_gaw import ReadGAW
 from .read_ghost import ReadGhost
 from .read_marcopolo import ReadMarcoPolo
-from .read_mscw_ctm import ReadMscwCtm
 
 if geopy_available() and coda_available():
     # the coda and geopy libraries are needed to read l2 data of the supported satellites
